+++ inst/python/sageMakerFunctions.py
@@ -16,7 +16,7 @@ def train_sagemaker(population, plpData, classifier, hyperParameters, container,
   sess = sagemaker.Session()
   
   y = population[:,1]
-  X = plpData[population[:,0],:]
+  X = plpData[population[:,0].astype(int),:]
   trainInds =population[:,population.shape[1]-1] >0
 
   print("Dataset has %s rows and %s columns" %(X.shape[0], X.shape[1]))
